+++ integration-test/testsrc/no/mnemonic/act/platform/test/integration/OriginIT.java
@@ -1,9 +1,15 @@
 package no.mnemonic.act.platform.test.integration;
 
+import no.mnemonic.act.platform.api.request.v1.CreateOriginRequest;
 import no.mnemonic.act.platform.dao.cassandra.entity.OriginEntity;
 import org.junit.Test;
 
+import javax.ws.rs.client.Entity;
+import javax.ws.rs.core.Response;
 import java.util.UUID;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 public class OriginIT extends AbstractIT {
 
@@ -23,6 +29,18 @@ public class OriginIT extends AbstractIT {
 
     // ... and check that it can be found via the REST API.
     fetchAndAssertList("/v1/origin", entity.getId());
+  }
+
+  @Test
+  public void testCreateOrigin() throws Exception {
+    // Create an Origin via the REST API ...
+    CreateOriginRequest request = new CreateOriginRequest()
+            .setName("origin");
+    Response response = request("/v1/origin").post(Entity.json(request));
+    assertEquals(201, response.getStatus());
+
+    // ... and check that it ends up in the database.
+    assertNotNull(getOriginManager().getOrigin(getIdFromModel(getPayload(response))));
   }
 
   private OriginEntity createOrigin() {