+++ CHANGELOG.md
@@ -7,3 +7,4 @@ The format is inspired by [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 ### Added
 - REST endpoint for retrieval of an Origin by ID, `GET /v1/origin/uuid/{id}`.
 - REST endpoint for listing Origins, `GET /v1/origin`.
+- REST endpoint for creating Origins, `POST /v1/origin`.